rename: util패키지에서 constant패키지로 파일 이동
