--- src/main/java/christmas/util/ValidatorConst.java
@@ -1,6 +1,0 @@
-package christmas.util;
-
-public class ValidatorConst {
-    public static final int COMPARE_EQUAL_VALUE_RESULT = 0;
-    public static final String ONLY_CATEGORY = "drink";
-}
