+++ src/main/java/christmas/util/constant/ValidatorConst.java
@@ -1,0 +1,6 @@
+package christmas.util.constant;
+
+public class ValidatorConst {
+    public static final int COMPARE_EQUAL_VALUE_RESULT = 0;
+    public static final String SINGLE_ORDER_CATEGORY = "drink";
+}
